TESTS: fixed builtin codegen (builtin*->builtin*l)

+++ tests/zbb/cpop/cpop64_manual.c
@@ -1,10 +1,9 @@
 #include <stdint.h>
 
-int test(int64_t rs) {
-    int count = 0;
+int64_t test(int64_t rs) {
+    int64_t count = 0;
     for (int i = 0; i < 64; ++i) {
         if (rs & (1ll << i)) ++count;
     }
     return count;
 }
-
